+++ src/lib/metadata-manager/index.ts
@@ -1,9 +1,14 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-import { ConnectionMembers } from "../connection/types/connection-members";
+import { Connection } from "../connection";
 import { CompassError } from "../error";
 import { CompassErrorCodeEnum } from "../error/types/error-code.enum";
-import { getEntitiesMetadata } from "./helpers/get-entities-metadata";
+import { setEntitiesMetadata } from "./helpers/set-entities-metadata";
+import { setSubEntitiesMetadata } from "./helpers/set-sub-entities-metadata";
+import {
+	convertEntityToDatabase,
+	ConvertEntityToDatabaseParams,
+} from "./methods/convert-entity-to-database";
 import { MetadataManagerEntities } from "./types/manager-metadata";
 
 export class MetadataManager<EntityExtraMetadata, ColumnExtraMetadata> {
@@ -25,7 +30,17 @@ export class MetadataManager<EntityExtraMetadata, ColumnExtraMetadata> {
 	private readonly entities: MetadataManagerEntities<
 		EntityExtraMetadata,
 		ColumnExtraMetadata
-	>;
+	> = {};
+
+	/**
+	 * Saves the metadata of all sub-entities
+	 *
+	 * Follows the same format that the entities property
+	 */
+	private readonly subEntities: MetadataManagerEntities<
+		EntityExtraMetadata,
+		ColumnExtraMetadata
+	> = {};
 
 	/**
 	 * ---------------------------------------------------
@@ -36,14 +51,23 @@ export class MetadataManager<EntityExtraMetadata, ColumnExtraMetadata> {
 	 */
 
 	public constructor(
-		connection: ConnectionMembers<EntityExtraMetadata, ColumnExtraMetadata>,
+		connection: Connection<EntityExtraMetadata, ColumnExtraMetadata>,
 	) {
-		this.entities = getEntitiesMetadata<
-			EntityExtraMetadata,
-			ColumnExtraMetadata
-		>({
+		setEntitiesMetadata<EntityExtraMetadata, ColumnExtraMetadata>({
+			entities: this.entities,
 			logger: connection.logger,
-			entities: connection.options.entities,
+			rawEntities: connection.options.entities,
+			connectionOptions: connection.options,
+		});
+
+		const allEntitiesColumns = Object.values(this.entities)
+			.map(entity => entity.columns)
+			.flat();
+
+		setSubEntitiesMetadata<EntityExtraMetadata, ColumnExtraMetadata>({
+			allEntitiesColumns,
+			subEntities: this.subEntities,
+			logger: connection.logger,
 			connectionOptions: connection.options,
 		});
 	}
@@ -58,6 +82,10 @@ export class MetadataManager<EntityExtraMetadata, ColumnExtraMetadata> {
 
 	public getAllEntitiesMetadata() {
 		return this.entities;
+	}
+
+	public getAllSubEntitiesMetadata() {
+		return this.subEntities;
 	}
 
 	public getAllEntitiesDatabaseNames() {
@@ -77,6 +105,21 @@ export class MetadataManager<EntityExtraMetadata, ColumnExtraMetadata> {
 		}
 
 		return entityMetadata;
+	}
+
+	public getSubEntityMetadata(subEntity: any) {
+		const subEntityMetadata = this.subEntities[subEntity.name];
+
+		if (!subEntityMetadata) {
+			throw new CompassError({
+				message: "SubEntity not Registered!",
+				code: CompassErrorCodeEnum.ENTITY_ERROR,
+				origin: "COMPASS",
+				details: ["SubEntity: ", subEntity],
+			});
+		}
+
+		return subEntityMetadata;
 	}
 
 	public getColumnMetadata(entity: any, columnName: string) {
@@ -100,5 +143,11 @@ export class MetadataManager<EntityExtraMetadata, ColumnExtraMetadata> {
 		return this.entities[entity.name]?.columns.filter(
 			columnMetadata => columnMetadata.primary,
 		);
+	}
+
+	public convertEntityToDatabase(
+		params: ConvertEntityToDatabaseParams,
+	): Record<string, any> {
+		return convertEntityToDatabase({ metadataManager: this }, params);
 	}
 }